Add initial methods for new transaction system
We plan to introduce a new transaction system which can protect db operations and also cover client notification. This is the first PR of the new system which aims to add new type and method for txn. Also create v2 methods for mutation and realtime transaction instead of updating those methods directly so we don't have to update all places in one run.